--- a/modules/models.py
+++ b/modules/models.py
@@ -183,9 +183,7 @@ def load_model(model_name):
         tokenizer = AutoTokenizer.from_pretrained(Path(f"{shared.args.model_dir}/{shared.model_name}/"))
     tokenizer.truncation_side = 'left'
 
-
     print(f"Loaded the model in {(time.time()-t0):.2f} seconds.")
-
     return model, tokenizer