This adds the option to replace llama's default attention with xformer's memory efficient attention.

After some testing, it does not appear that xformers decreases llama's memory footprint all that much nor does it speed up regular inference as much as it does with stable diffusion. However, it does increase the inference times for when you generate with a large context.

Normal Attention
![Capture1](https://user-images.githubusercontent.com/31809330/230763784-ff573600-ce6f-41cc-af63-b14a0a912488.PNG)

Xformers Attention
![Capture2](https://user-images.githubusercontent.com/31809330/230763787-ec73aeb4-44c6-4772-8c87-a1bbff9dfbd1.PNG)