@thot-experiment 2023-04-09
Looks like great work! However, it seems like judging by your screenshot you may have meant to say that it **decreases** inference time? 

> However, it does **increase** the inference times for when you generate with a large context.

Recently IIRC torch2 has implemented this or something similar natively without the need for xformers? I don't know much about the nitty gritty of it so I'm not even sure why I'm bringing it up, but maybe that could be a way further forward?