@Ph0rk0z 2023-04-09
Why only llama? Why not others? I will definitely give it a whirl... I need all the help I can get on my old card. Did it work with GPTQ?