@EyeDeck 2023-04-09
Some large-context test runs (WSL2, GPTQ-for-LLaMA triton branch, RTX 3090, LLaMA 30B 4-bit act-order groupsize 128):

Control:
>Output generated in 65.74 seconds (3.03 tokens/s, 199 tokens, context 1848)
Output generated in 69.57 seconds (2.86 tokens/s, 199 tokens, context 1848)
Output generated in 63.45 seconds (3.14 tokens/s, 199 tokens, context 1848)
Output generated in 64.79 seconds (3.07 tokens/s, 199 tokens, context 1848)
Output generated in 60.89 seconds (3.27 tokens/s, 199 tokens, context 1848)
Output generated in 62.67 seconds (3.18 tokens/s, 199 tokens, context 1848)
Output generated in 64.70 seconds (3.08 tokens/s, 199 tokens, context 1848)
Output generated in 60.99 seconds (3.26 tokens/s, 199 tokens, context 1848)

Average: 3.105 tokens/s

---

--sdp-attention:
>Output generated in 49.33 seconds (4.03 tokens/s, 199 tokens, context 1848)
Output generated in 55.25 seconds (3.60 tokens/s, 199 tokens, context 1848)
Output generated in 61.90 seconds (3.21 tokens/s, 199 tokens, context 1848)
Output generated in 63.69 seconds (3.12 tokens/s, 199 tokens, context 1848)
Output generated in 63.70 seconds (3.12 tokens/s, 199 tokens, context 1848)
Output generated in 61.64 seconds (3.23 tokens/s, 199 tokens, context 1848)
Output generated in 71.64 seconds (2.78 tokens/s, 199 tokens, context 1848)
Output generated in 61.95 seconds (3.21 tokens/s, 199 tokens, context 1848)

Average: 3.255 tokens/s

---

--xformers:
>Output generated in 55.00 seconds (3.62 tokens/s, 199 tokens, context 1848)
Output generated in 61.08 seconds (3.26 tokens/s, 199 tokens, context 1848)
Output generated in 59.70 seconds (3.33 tokens/s, 199 tokens, context 1848)
Output generated in 57.36 seconds (3.47 tokens/s, 199 tokens, context 1848)
Output generated in 63.45 seconds (3.14 tokens/s, 199 tokens, context 1848)
Output generated in 50.22 seconds (3.96 tokens/s, 199 tokens, context 1848)
Output generated in 56.78 seconds (3.50 tokens/s, 199 tokens, context 1848)
Output generated in 56.92 seconds (3.50 tokens/s, 199 tokens, context 1848)

Average: 3.457 tokens/s

---

Can't measure any VRAM usage difference, all three climb up to 99.8-100% as measured by HWiNFO.
At least in this test, --xformers is about 11% faster than baseline, so it's an improvement.