@oobabooga 2023-04-10
That's very impressive. In my tests, there was a decrease in performance with xformers using LLaMA 7b in 16-bit mode:

* Without xformers: Output generated in 6.00 seconds (33.15 tokens/s, 199 tokens, context 16)
* With xformers: Output generated in 6.61 seconds (30.10 tokens/s, 199 tokens, context 16)

But there are many other possibilities to experiment with (8-bit, 4-bit, offloading). 

Thank you for submitting such sophisticated PR.