@VldmrB 2023-04-10
I was all excited, seeing as building xformers from source and using it in stable-web-ui increased generation speed by 100%+ for me, but here, after building it from scratch again, it's about the same, probably slightly slower.

I'm on Windows, but that probably doesn't matter.
For those curious, here's the excellent post that describes building xformers from source, on Windows: https://github.com/AUTOMATIC1111/stable-diffusion-webui/issues/2449#issuecomment-1404540735
Maybe there's something that's off and it will be fixed, fingers crossed.