@EyeDeck 2023-04-10
I think it's more just that the workload doesn't happen to benefit as much from xformers as some others do. The GPTQ-for-LLaMA guy, qwopqwop200, mentioned that he tried implementing xformers over there once, and he didn't see any benefit either, so he gave up on that code.