@Ph0rk0z 2023-04-10
For me it's the same. It didn't really do much one way or another. I had to upgrade pytorch to use it.



```
xformers

7b
Output generated in 3.37 seconds (3.56 tokens/s, 12 tokens, context 207)
Output generated in 2.97 seconds (4.04 tokens/s, 12 tokens, context 207)
Output generated in 2.95 seconds (4.07 tokens/s, 12 tokens, context 207)

7b fp16
Output generated in 2.59 seconds (5.78 tokens/s, 15 tokens, context 418)
Output generated in 2.28 seconds (6.59 tokens/s, 15 tokens, context 418)
Output generated in 2.30 seconds (6.52 tokens/s, 15 tokens, context 418)


30b
Output generated in 8.22 seconds (0.85 tokens/s, 7 tokens, context 207)
Output generated in 7.94 seconds (0.88 tokens/s, 7 tokens, context 207)
Output generated in 7.93 seconds (0.88 tokens/s, 7 tokens, context 207)

Output generated in 31.35 seconds (1.08 tokens/s, 34 tokens, context 418)
Output generated in 16.94 seconds (0.94 tokens/s, 16 tokens, context 418)
Output generated in 19.26 seconds (0.99 tokens/s, 19 tokens, context 418)

30b gptqv2
Output generated in 9.34 seconds (1.93 tokens/s, 18 tokens, context 418)
Output generated in 9.02 seconds (2.00 tokens/s, 18 tokens, context 418)
Output generated in 9.07 seconds (1.98 tokens/s, 18 tokens, context 418



non xformers for GPTQv1 
7b
Output generated in 3.24 seconds (3.71 tokens/s, 12 tokens, context 207)
Output generated in 2.96 seconds (4.05 tokens/s, 12 tokens, context 207)
Output generated in 2.97 seconds (4.04 tokens/s, 12 tokens, context 207)

7b fp16
Output generated in 2.50 seconds (6.01 tokens/s, 15 tokens, context 418)
Output generated in 2.20 seconds (6.83 tokens/s, 15 tokens, context 418)
Output generated in 2.22 seconds (6.77 tokens/s, 15 tokens, context 418)

30b
Output generated in 8.19 seconds (0.85 tokens/s, 7 tokens, context 207)
Output generated in 7.90 seconds (0.89 tokens/s, 7 tokens, context 207)
Output generated in 7.86 seconds (0.89 tokens/s, 7 tokens, context 207)

Output generated in 12.55 seconds (0.80 tokens/s, 10 tokens, context 418)
Output generated in 17.66 seconds (0.96 tokens/s, 17 tokens, context 418)
Output generated in 24.81 seconds (1.05 tokens/s, 26 tokens, context 418)

30b gptqv2
Output generated in 9.13 seconds (1.97 tokens/s, 18 tokens, context 418)
Output generated in 8.84 seconds (2.04 tokens/s, 18 tokens, context 418)
Output generated in 8.88 seconds (2.03 tokens/s, 18 tokens, context 418)

```