@dronelove 2023-04-10
can someone explain how to enable xformers? I tried adding --xformers to startwebui.bat but got an error saying "xformers is not defined." thanx