@thot-experiment 2023-04-10
> I tried adding --xformers to startwebui.bat but got an error saying "xformers is not defined." 
I have the same error, i've confirmed that xformers is installed inside umamba via `pip list` which returns `xformers               0.0.18`

additionally creating a file `test.py` with one line `import xformers` returns the same error, so this is perhaps some sort of a python issue (not my native language so I'm not sure where to go from here)  `which python` returns the correct path `ooba/installer_files/env/python` while inside umamba cmd